requirements.txt
pandas==2.1.4
biopython
torch
beautifulsoup4
transformers==4.52.1
qwen_vl_utils
accelerate
datasets
esm
requests
biotite
pymupdf4llm
pymupdf
torch
torchvision
layoutparser
SQLAlchemy==1.4.54
git+https://github.com/facebookresearch/detectron2.git@ff53992b1985b63bd3262b5a36167098e3dada02
git+https://github.com/celalp/usearch-molecules.git@0a64c027ba8ebc0a60270287a4d6512e7cd77836
chonkie[st]
chonkie[model2vec]
pysam
lxml
pgvector
adapters
colpali-engine>=0.3.0
pytest
docker
openpyxl
model2vec